requirements.txt
asgiref==3.6.0
autopep8==2.0.2
dj-database-url==1.2.0
Django==4.1.7
django-js-asset==2.0.0
django-mptt==0.14.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
Pillow==9.4.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
python-dotenv==1.0.0
sqlparse==0.4.3
tomli==2.0.1
